src/rest_api/flask_app/requirements.txt
cassandra-driver==3.25.0
Flask==2.0.1
Flask-RESTful==0.3.9
jsonschema==3.2.0
numpy==1.19.5
pandas==1.1.5
requests==2.25.1
PyJWT==2.1.0
passlib==1.7.4
gunicorn==20.1.0
python-dotenv==0.17.1
python-logstash-async==2.3.0
Werkzeug==2.2.2
flasgger==0.9.7.1

